chore: apply PR comments
